--- a/Sources/tart/OCI/Layerizer/DiskV2.swift
+++ b/Sources/tart/OCI/Layerizer/DiskV2.swift
@@ -10,7 +10,7 @@ class DiskV2: Disk {
   // A zero chunk for faster than byte-by-byte comparisons
   //
   // Assumes that the other Data(...) is equal in size, but it's fine to get a false-negative
-  // on the last block since it costs only 64 KiB of excess data per 512 MiB layer.
+  // on the last block since it costs only 4 MiB of excess data per 512 MiB layer.
   //
   // Some simple benchmarks ("sync && sudo purge" command was used to negate the disk caching effects):
   // +--------------------------------------+---------------------------------------------------+
@@ -19,7 +19,7 @@ class DiskV2: Disk {
   // | Data(...) == zeroChunk               | 2.16s user 11.71s system 73% cpu 18.928 total     |
   // | Data(...).contains(where: {$0 != 0}) | 603.68s user 12.97s system 99% cpu 10:22.85 total |
   // +--------------------------------------+---------------------------------------------------+
-  private static let holeGranularityBytes = 64 * 1024
+  private static let holeGranularityBytes = 4 * 1024 * 1024
   private static let zeroChunk = Data(count: holeGranularityBytes)
 
   static func push(diskURL: URL, registry: Registry, chunkSizeMb: Int, concurrency: UInt, progress: Progress) async throws -> [OCIManifestLayer] {