--- a/Sources/tart/VM.swift
+++ b/Sources/tart/VM.swift
@@ -1,6 +1,5 @@
 import Foundation
 import Virtualization
-import AsyncAlgorithms
 import Semaphore
 
 struct UnsupportedRestoreImageError: Error {