@edigaryev 2024-12-11
> Will post the [profiling results](https://github.com/cirruslabs/tart/blob/main/PROFILING.md#using-xctrace1) tomorrow.

Just to clarify, the main goal I see here in doing profiling is to avoid regressions, namely excessive allocations.

The goal is not to measure the performance, as this is already done in the benchmarks above.

I've also excluded `--concurrency 8` from profiling as it's insanely slow and because `--concurrency 4` should already highlight any issues, if any.

## `main` branch of Tart

### `--concurrency 1`

<img width="1143" alt="profiler-allocations-target-tart-concurrency-1" src="https://github.com/user-attachments/assets/e1f2b5c3-6c9c-40ea-8028-b08ccc457eae">

### `--concurrency 2`

<img width="1143" alt="profiler-allocations-target-tart-concurrency-2" src="https://github.com/user-attachments/assets/cf5a591e-22c2-4018-bbb8-2ebda243ec59">

### `--concurrency 4`

<img width="1143" alt="profiler-allocations-target-tart-concurrency-4" src="https://github.com/user-attachments/assets/086870d7-91c8-4363-8b18-57ab52efb8db">

## `main` branch of Tart with 4 MiB hole granularity size

### `--concurrency 1`

<img width="1143" alt="profiler-allocations-target-tart-main-branch-with-4mib-holes-concurrency-1" src="https://github.com/user-attachments/assets/b553f8c7-7be6-4e9b-8ae6-687cbd6cd391">

### `--concurrency 2`

<img width="1143" alt="profiler-allocations-target-tart-main-branch-with-4mib-holes-concurrency-2" src="https://github.com/user-attachments/assets/046b7beb-1a66-4e66-a383-d4d3ea8ab4b1">

### `--concurrency 4`

<img width="1143" alt="profiler-allocations-target-tart-main-branch-with-4mib-holes-concurrency-4" src="https://github.com/user-attachments/assets/53b45ce6-9d41-4d9b-b6eb-8de81bf33be1">

## This PR

### `--concurrency 1`

<img width="1143" alt="profiler-allocations-target-this-pr-concurrency-1" src="https://github.com/user-attachments/assets/ef8387d0-849b-4fd6-8ad5-cfd8c592fbf7">

### `--concurrency 2`

<img width="1143" alt="profiler-allocations-target-this-pr-concurrency-2" src="https://github.com/user-attachments/assets/39433d0a-b52e-438c-a352-ee69011e92f7">

### `--concurrency 4`

<img width="1143" alt="profiler-allocations-target-this-pr-concurrency-4" src="https://github.com/user-attachments/assets/cb366666-48d2-4b88-902e-82fd93a78307">